feat: aborting first approach
having a single sorted array is nice but fails for the edge case when
the last asteroids are in line, i.e. same phase. In this case we would
need a reset indicator to vaporize two asteroids having same phase,
which makes the whole approach too complicated (skip already vaporized
ones, skip same phase, but don't skip if {{weird conditions}}.